requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
email_validator==2.1.0.post1
pandas
scikit-learn
joblib
numpy
pymysql
python-dotenv
opencv-python
reportlab==4.0.0
Flask-JWT-Extended==4.5.3
Pillow
